refactor: adjust chart code dir
