+++ package.json
@@ -25,11 +25,11 @@
     "pinia": "^2.0.11",
     "qrcode.vue": "^3.2.2",
     "tdesign-icons-vue-next": "^0.0.6",
-    "tdesign-vue-next": "~0.11.2",
+    "tdesign-vue-next": "^0.13.0",
     "tvision-color": "^1.3.1",
     "vue": "^3.2.31",
-    "vue-router": "^4.0.11",
-    "vue-clipboard3": "^2.0.0"
+    "vue-clipboard3": "^2.0.0",
+    "vue-router": "^4.0.11"
   },
   "devDependencies": {
     "@commitlint/cli": "^16.2.1",
